Build three pages out of Laravel
This is a demo of Laravel application by building Home, About and
Contact pages.  At the top of each page, navigational links connect to
each of these pages.

Use Laravel component for layout of the common parts of all three pages.

$ sail artisan make:component --view -- layout

  INFO  Component created successfully.

$ ls resources/views/components/
layout.blade.php

Use $slot in the layout for the unique part of each page.

Additional work is needed for dynamic page title.