Add view component for navlink
Build a view component of an anchor tag which dynamically serves the
navlink for reusability.

This part <a {{ $attributes }}> is the opening anchor tag.

This part {{ $attributes }} is Blade syntax specific to Laravel views.
It inserts the value of the PHP variable $attributes in the opening tag.

This variable $attributes is an array containing key-value pairs, where:

- The key represents the HTML attribute name (e.g., "href", "class",
  "id").
- The value represents the corresponding attribute value for the key.