Update pages with Tailwind stacked layouts
First obtain Tailwind stacked layouts by copying the free markup as a
template to update the layout.  Use Home, About and Contact as the nav
links in the template.

https://tailwindui.com/components/application-ui/application-shells/stacked

For the logo, use my own icon and 'Laravel in Docker' is a suitable
string for the alt text.  Use GitHub octocat to replace Tailwind
author's headshot.  Also use my own name and email address.  These
changes cover both desktop and mobile.

In the layout, add heading slot so that the name of heading can be used
by Home, About and Contact pages for their unique values.  Also add full
height and gray color for background.