Let link be highlighted for the current page only
Home, About and Contact are the links for navigation.  Tailwind makes
Home page link always highlighted whether it is the current page or not.
Add the feature that makes the link highlighted when it is the current
page only.

The use of request()->is('/') is shorthand of these two lines:

use Illuminate\Support\Facades\Request;
Request::is('/');

https://laravel.com/api/11.x/Illuminate/Support/Facades/Request.html#method_is

The static class attribute is now dynamic with request()->is() PHP
expression embedded in Blade templating {{..}}.