Use navlink component for hard-coded anchor tag
Home, About and Contact links are anchor tags hard-coded with almost
identical attributes.  Use navlink component to reduce the repetitive
attributes as was done before the introduction of tailwind CSS.  Apply
navlink component to both desktop and mobile.

All anchor tags were hard-coded.  This had many issues, including Home
link always highlighted even when it was not the current page.  Fix this
by the use of boolean value to determine if the link is the current page
or not.

Use :active="foo" to retain the boolean value of foo.  Without ':' only
string is assigned to 'active' prop, not the boolean value.  The ':'
prefix informs Blade that the prop assignment is a PHP expression.

In Blade components, props are introduced by @props() directive and the
directive must be the first line.

@props([ 'foo' => $bar ])