Add jobs page
The job title and the salary are the keys of an associative array.  A
list of these arrays form an indexed array.  This indexed array becomes
the value of the outer associative array with the key by the name of
jobs.  The outer associative array, or the whole array, is passed in as
data to view() function.  Nevertheless, the key of jobs is used as $jobs
to refer to the indexed array.

Build Jobs page using layout as Home page.  Loop through the indexed
array by $jobs for the values of job title and salary.

Update layout for Jobs page.  Remove About page.

https://laravel.com/api/11.x/[Global_Namespace].html#function_view