Add individual Job page
First add the route for /jobs/{id}, where {id} is determined by the job
link the user clicks.  \Illuminate\Support\Arr has first() method which
can pick up the right array if {id} has a match.

Add the individual Job page that displays the array with info of job
title and salary.  The array is found in the previous step.  If no array
is found, for example, manually entering the URL of jobs/34, the page
breaks.

Update Jobs page with the new info of id along with job title and salary
to display the matched array.