Add Job class to encapsulate array
Declaring $jobs array outside a class is an example for teaching only.
OOP provides encapsulation for local object.  In a better sense, $jobs
array will be encapsulated in a class.  In production, $jobs will be
populated by querying the database via a method.

The use of 'use' keyword serves the same teaching purpose.  When class
is used, 'use' keyword in this setting is not needed any more.  It's
quite distinctive to use scope resolution operator '::' for access of
class properties and methods.