Add find() method to Job model
The logic of finding the matching array should be executed in Models,
not in Controllers.  Move the matching mechanism from Controllers to
Models.  This also makes the code cleaner.

Use find($id) in lieu of find(int $id).  The passed in value to
Job::find() could be integer string, couldn't it?