Use first_name and last_name for UserFactory
In User model, users database table is updated with first_name and
last_name in lieu of name column.  When factory is used, UserFactory
must match the columns of users database table for the records to be
inserted.

Verify a big number of records is inserted successfully.

> use App\Models\User;
> User::factory(100)->create();
> User::all();

Verify TablePlus shows the same record insertion.

Before this patch, this error would show.

> User::factory()->create();

   Illuminate\Database\QueryException  SQLSTATE[42S22]: Column not
   found: 1054 Unknown column 'name' in 'field list' (Connection: mysql,
   SQL: insert into `users` (`name`, `email`, `email_verified_at`,
   `password`, `remember_token`, `updated_at`, `created_at`) values
   (Cydney Ondricka, meaghan.kuhlman@example.net, 2024-05-28 04:55:53,
   $2y$12$8.VBDfTk9oPtKqFzAeb1y.yvMMBWLm5dtffcTM7K7XJIKsV.OJkWG,
   8UwIOHuiRN, 2024-05-28 04:55:53, 2024-05-28 04:55:53)).