Add admin state to UserFactory
Add admin to userFactory definition and set its default value to false.

Use the state facility for admin().  When factory() is executed through
admin(), the value of admin in the database will be set to true in lieu
of default false.

Reset User database and verify admin value is created correctly.

> User::factory()->create();
> User::factory()->admin()->create();
> User::find(1)->admin;
= 0
> User::find(2)->admin;
= 1

Use boolean() in lieu of tinyInteger() for efficiency.