Add title and content to posts table and PostFactory
For Post model, we first set title and content in posts table; and
accordingly, we also set these two attributes in post factory to match
the table.

The title is different from name, and we will use ->sentence() for the
post title.  The content is more than text and we will use ->paragraph()
for the post content.