Add job_tag pivot table
A job can have many tags and a tag can be assigned to many jobs.

Between Job and Tag, the relationship is many-to-many.  To make this
relationship work, we need to have a pivot table to regulate.  The
job_tag table is created for this purpose.

Usually job_tag migration will be added by its own migration table.  In
this case we add it inside tags migration table.

Use foreignIdFor() method to generate both job_listing_id and tag_id.
Specify job_listing_id in lieu of default job_id to show the connection
to job_listings table.

Add drop schema for job_tag if we need to drop this table.