Prevent lazy loading
By default Laravel app has lazy loading on, which could introduce N+1
problem.  In our case, jobs are queried individually through each
employer and this will lead to performance issues.  We should prevent
lazy loading when the app starts.

With lazy loading turned off at the beginning of the app, we can't use
lazy loading later in the app, but through the use of with().

We have to be aware if the number of jobs is big, we will have to use
pagination to circumvent the possible bottleneck.  Thus, we will strike
a good balance between performance and potential issues.