Use simple pagination
With simplePaginate() method call, the pagination has only Previous and
Next buttons for navigation, but without individual page numbers.  Add
this call to the tree and see the effects for comparison with others.
In a simple layout, it might be preferable.

With get(), querying job_listings table has no limit.  Using paginate()
or simplePaginate(), querying job_listings table has limit of 3 or 4
respectively, in addition to the offset.