Remove obsolete Job views
The jobs and single job views have been migrated to jobs directory and
their old counterparts are now obsolete.  Tidy up resources/views/.

This concludes the migration of jobs views.  The series of commits
dbf0f724d770..6f67c71451e4 presents the steps of views scaffolding for
jobs of Job model.  It might be extended for other models as well.