Add user input validation
For request(), it gets an instance of the current request or an input
item from the request.

->validate() takes an associative array which has title and salary keys.
Their respective value is an indexed array of a list of requirements.

If validation fails, the application stops execution there and is
redirected.  Error messages are saved in $errors.

Next, will use $errors to display the error messages in Blade.