Add Edit Job page
Add the link to Edit Job page in each individual job post.

Add the route to Edit Job page.

Add the form to Edit Job page using Create Job page as a template.  At
this stage of building, exclude form method and action.

Use the old value of input field for the value.