Use implicit route model binding
Route model binding in Laravel allows to automatically inject instances
of models directly into routes or controller actions.  Instead of
manually retrieving the model instance from the database, Laravel will
do it automatically, simplifying the code and making it more readable.

With implicit binding, Laravel will automatically resolve the model
based on the route parameter.

Thus, it automatically resolves the {job} parameter to an instance of
Job model by looking for {job} as {id} in the database.