Use Route::resource() to manage jobs
Route::resource() is a method used to create a set of routes for a
resourceful controller in a single line of code. This method
automatically creates standard routes that follow RESTful conventions
for common actions such as creating, reading, updating, and deleting
(CRUD) resources, in our example, jobs.

This is the standard implementation.

Route::resource('resources', ResourceController::class);

1) GET /jobs: index method to list all jobs

2) GET /jobs/create: create method to show the form for creating a new
   job

3) POST /jobs: store method to store a new job

4) GET /jobs/{job}: show method to display a specific job

5) GET /jobs/{job}/edit: edit method to show the form for editing a job

6) PATCH/PUT /jobs/{job}: update method to update a specific job

7) DELETE /jobs/{job}: destroy method to delete a specific job