Update /posted mail route
Commit f7e65b17b3c2 ("Use newly created job for JobPosted class")
updated the constructor of JobPosted class.  Subsequently the
instantiation of JobPosted class must be updated accordingly.  Use the
newly created job reference for instantiation.

Job::with('employer') eagerly loads the associated Employer model for
each Job instance retrieved from the database.

Job::with(): This is a Laravel Eloquent method used for eager loading
relationships.  It specifies which related models should be loaded along
with the primary model (in this case, Job) when querying the database.

'employer': This is the name of the relationship defined in the Job
model that links it to the Employer model.

Thus, these are the steps that take place.

1) Fetch the latest Job record from the database.

2) Simultaneously fetch related Employer record for the retrieved Job.

3) Populate the employer relationship on Job object with corresponding
   Employer instance.