Put newly created job in queue
Use queue() to put the newly created job in queue rather than send it
out.  Now the sending part will be executed in the background, which
requires the worker.

$ sail artisan queue:work

   INFO  Processing jobs from the [default] queue.

  2024-07-21 03:31:03 App\Mail\JobPosted ...................  RUNNING
  2024-07-21 03:31:04 App\Mail\JobPosted ...................  1s DONE

APP_URL defaults to http://localhost when sail artisan is run, so it
needs to be reset.  Update .env accordingly.

APP_URL is documented in config/app.php on how it should be used when
the worker is executed in the console.

  /*
  |---------------------------------------------------------------------
  | Application URL
  |---------------------------------------------------------------------
  |
  | This URL is used by the console to properly generate URLs when using
  | the Artisan command line tool. You should set this to the root of
  | the application so that it's available within Artisan commands.
  |
  */

  'url' => env('APP_URL', 'http://localhost'),