Use POST method for link in Nav vue
By default Inertia uses GET for the anchor tag in vue pages.  When POST
method is specified, Inertia will use the specified method of POST in
the event handler of the link.

POST method is usually used in forms, but as with Inertia it can be used
outside of forms.