Turn /spa/logout into a button
The link to /spa/logout is an anchor tag with POST method.  If the link
opens a blank tab, for example, Ctrl + Click, GET will be used instead.
This will trigger the exception documented in commit a24879ebaa4e ("Add
/spa/logout link in Nav vue").

To get around this, make /spa/logout link a button.  It is still a link,
but it won't open a blank tab any more.  Thus, POST method is preserved.