Add shared object to Users
Use HandleInertiaRequests middleware and turn the current time into a
shared object.  Use this shared object first in Users vue page.  Use the
name of 'clock' to minimize the conflict with the name of the time
string in Home vue page.

Add the link to update the current time in Users as well.