Use the name of time for current time
This is a cleanup patch.  With the name conflict gone, use the name of
time for the current time object.

This series changes from the use of time string in the routes to the use
of time object in share() method of Inertia middleware.  It not only
makes the routes cleaner, gets rid of redundant props, but more
importantly it demonstrates the power of middleware offered by Inertia.
Through middleware, sharing of objects becomes an easy task in multiple
vue pages.  The use of middleware is the way to go.