Add test of 'it belongs to an employer'
Add the test that the job created belongs to the employer.  Follow AAA
procedure of Arrange, Act and Assert.  Particular to this test, combine
Act and Assert due to the simple nature of the test.

Use the alias of it in lieu of test to construct the test.

By default pest is used for feature test only.  Add unit test to pest.
Without the unit test added, it will throw this exception.

   FAILED  Tests\Unit\JobTest > it belongs to an employer
   RuntimeException
 A facade root has not been set.

While there, add RefreshDatabase so that the test is run in pristine
state.