Add the title in Inertia template
The title in Inertia template has higher precedence than that in Home,
Users and Settings vue pages.  The template title combined with vue page
title in the end forms the resulting title.  If vue page title is not
set, the template title would be used as its default value.

Move process under setup and add space to improve template readability.