Filter users collection to display
Commit 1073321c59ab ("Let Inertia display all database users") imposes a
security issue that could potentially display sensitive info.  Inertia,
unlike views, is for front end and will display anything handed to it.

To circumvent the issue, filter the collection first and then hand it to
Inertia.  Thus, use map() method in the controller to filter out all the
info except for the first name and last name.